feat: finish lab-2 skeleton

+++ Shurshavchiki/BusinessLogic/src/main/java/ru/shurshavchiki/businessLogic/colorSpace/channelChoosers/ChannelChooser.java
@@ -2,6 +2,6 @@ package ru.shurshavchiki.businessLogic.colorSpace.channelChoosers;
 
 public interface ChannelChooser {
     String getMagicNumber();
-    byte[] apply(byte[] allChannelsData);
-    byte[] fillAllChannels(byte[] rawData);
+    float[] apply(float[] allChannelsData);
+    float[] fillAllChannels(float[] rawData);
 }
